# server/ui/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenAPI 2.0 documentation for [Mailpit](https://github.com/axllent/mailpit).",
    "title": "Mailpit API",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/axllent/mailpit"
    },
    "license": {
      "name": "MIT license",
      "url": "https://github.com/axllent/mailpit/blob/develop/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/message/{ID}/part/{PartID}/thumb": {
      "get": {
        "description": "This will return a cropped 180x120 JPEG thumbnail of an image attachment.\nIf the image is smaller than 180x120 then the image is padded. If the attachment is not an image then a blank image is returned.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "image/jpeg"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get an attachment image thumbnail",
        "operationId": "ThumbnailParams",
        "parameters": [
          {
            "type": "string",
            "example": "4oRBnPtCXgAqZniRhzLNmS",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "example": "2",
            "description": "Attachment part ID",
            "name": "PartID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/BinaryResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/v1/tags": {
      "get": {
        "description": "Returns a JSON array of all unique message tags.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Get all current tags",
        "operationId": "GetAllTags",
        "responses": {
          "200": {
            "$ref": "#/responses/ArrayResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "description": "This will overwrite any existing tags for selected message database IDs. To remove all tags from a message, pass an empty tags array.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Set message tags",
        "operationId": "SetTagsParams",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "Tags",
                "IDs"
              ],
              "properties": {
                "IDs": {
                  "description": "Array of message database IDs",
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "4oRBnPtCXgAqZniRhzLNmS",
                    "hXayS6wnCgNnt6aFTvmOF6"
                  ]
                },
                "Tags": {
                  "description": "Array of tag names to set",
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "Tag 1",
                    "Tag 2"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/tags/{Tag}": {
      "put": {
        "description": "Renames an existing tag.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Rename a tag",
        "operationId": "RenameTagParams",
        "parameters": [
          {
            "type": "string",
            "example": "Old name",
            "description": "The url-encoded tag name to rename",
            "name": "Tag",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "Name"
              ],
              "properties": {
                "Name": {
                  "description": "New name",
                  "type": "string",
                  "example": "New name"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "Deletes a tag. This will not delete any messages with the tag, but will remove the tag from any messages containing the tag.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Delete a tag",
        "operationId": "DeleteTagParams",
        "parameters": [
          {
            "type": "string",
            "example": "My tag",
            "description": "The url-encoded tag name to delete",
            "name": "Tag",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/view/{ID}.html": {
      "get": {
        "description": "Renders just the message's HTML part which can be used for UI integration testing.\nAttached inline images are modified to link to the API provided they exist.\nNote that is the message does not contain a HTML part then an 404 error is returned.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "text/html"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "testing"
        ],
        "summary": "Render message HTML part",
        "operationId": "GetMessageHTMLParams",
        "parameters": [
          {
            "type": "string",
            "example": "4oRBnPtCXgAqZniRhzLNmS",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/HTMLResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/view/{ID}.txt": {
      "get": {
        "description": "Renders just the message's text part which can be used for UI integration testing.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "testing"
        ],
        "summary": "Render message text part",
        "operationId": "GetMessageTextParams",
        "parameters": [
          {
            "type": "string",
            "example": "4oRBnPtCXgAqZniRhzLNmS",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TextResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    }
  },
...
  "responses": {
...
    "ArrayResponse": {
      "description": "Plain JSON array response",
      "schema": {
        "type": "array",
        "items": {
          "type": "string"
        }
      }
    },
    "BinaryResponse": {
      "description": "Binary data response which inherits the attachment's content type.",
      "schema": {
        "type": "string"
      }
    },
    "ErrorResponse": {
      "description": "Server error will return with a 400 status code\nwith the error message in the body",
      "schema": {
        "type": "string"
      }
    },
    "HTMLResponse": {
      "description": "HTML response",
      "schema": {
        "type": "string"
      }
    },
...
    "NotFoundResponse": {
      "description": "Not found error will return a 404 status code",
      "schema": {
        "type": "string"
      }
    },
    "OKResponse": {
      "description": "Plain text \"ok\" response",
      "schema": {
        "type": "string"
      }
    },
    "TextResponse": {
      "description": "Plain text response",
      "schema": {
        "type": "string"
      }
    },
...
  }
}